Make abstract async security handler
